# docs/openapi.json
{
  "components": {
    "schemas": {
      "Draft": {
        "properties": {
          "body": {
            "description": "free text, up to 4096 bytes",
            "type": "string"
          },
          "title": {
            "description": "short label, 1-200 bytes after trimming",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "type": "object"
      },
      "Error": {
        "properties": {
          "error": {
            "description": "human-readable reason",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "Note": {
        "properties": {
          "body": {
            "description": "free text, up to 4096 bytes",
            "type": "string"
          },
          "created_at": {
            "description": "UTC creation timestamp",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "server-assigned UUID",
            "type": "string"
          },
          "title": {
            "description": "short label, 1-200 bytes after trimming",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "body",
          "created_at"
        ],
        "type": "object"
      },
      "Status": {
        "properties": {
          "status": {
            "description": "probe result",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
      "Version": {
        "properties": {
          "commit": {
            "description": "git commit, with -dirty when built from a dirty tree",
            "type": "string"
          },
          "version": {
            "description": "release version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "commit"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API over the notes domain. Generated by `devctl openapi` — never hand-edited.",
    "title": "go-agents API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/healthz": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            },
            "description": "the service is alive"
          }
        },
        "summary": "Liveness probe. Always 200 while the process is running."
      }
    },
    "/readyz": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            },
            "description": "ready to serve"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "a dependency is unavailable"
          }
        },
        "summary": "Readiness probe. Touches the store, so a broken dependency shows here."
      }
    },
    "/v1/notes": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Note"
                  },
                  "type": "array"
                }
              }
            },
            "description": "the notes"
          }
        },
        "summary": "List every note, newest first."
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Draft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            },
            "description": "the created note"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "missing or bad token"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a note."
      }
    },
    "/v1/notes/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "deleted"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "missing or bad token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "no note with that ID"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete one note by ID."
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            },
            "description": "the note"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "no note with that ID"
          }
        },
        "summary": "Fetch one note by ID."
      }
    },
    "/version": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Version"
                }
              }
            },
            "description": "version and commit"
          }
        },
        "summary": "Build provenance of the running binary."
      }
    }
  }
}